requirements.txt
tensorflow==2.15.0
numpy==1.24.3
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
Pillow==10.1.0
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
pandas==2.1.4
fastapi==0.109.0
uvicorn==0.25.0
python-multipart==0.0.6
pydantic==2.5.3
albumentations==1.3.1
imgaug==0.4.0
tqdm==4.66.1
tensorboard==2.15.1
pyyaml==6.0.1
jsonschema==4.20.0
scipy==1.11.4
requests==2.31.0

# MLOps Dependencies
mlflow==2.9.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
flake8==6.1.0
black==23.12.1
mypy==1.7.1

